Rename R class import coming from design system module
From
com.hedvig.android.core.designsystem.R
to
com.hedvig.android.core.design.system.R

According to the new auto-generated name
