+++ app/app/src/main/kotlin/com/hedvig/app/feature/profile/ui/tab/ProfileDestination.kt
@@ -209,10 +209,10 @@ private fun ColumnScope.ProfileItemRows(
       iconPainter = when {
         profileUiState.euroBonus == null -> ColorPainter(Color.Transparent)
         profileUiState.euroBonus.code != null -> {
-          painterResource(com.hedvig.android.core.designsystem.R.drawable.ic_checkmark_in_circle)
+          painterResource(com.hedvig.android.core.design.system.R.drawable.ic_checkmark_in_circle)
         }
 
-        else -> painterResource(com.hedvig.android.core.designsystem.R.drawable.ic_info)
+        else -> painterResource(com.hedvig.android.core.design.system.R.drawable.ic_info)
       },
       onClick = navigateToEurobonus,
     )
