+++ app/app/src/main/kotlin/com/hedvig/app/feature/referrals/ui/redeemcode/RedeemCodeBottomSheet.kt
@@ -86,7 +86,7 @@ abstract class RedeemCodeBottomSheet : BottomSheetDialogFragment() {
 
   private fun wrongPromotionCode(errorMessage: String) {
     binding.textField.errorIconDrawable =
-      requireContext().compatDrawable(com.hedvig.android.core.designsystem.R.drawable.ic_warning_triangle)
+      requireContext().compatDrawable(com.hedvig.android.core.design.system.R.drawable.ic_warning_triangle)
     binding.textField.error = errorMessage
   }
 }
